refactor(oauth): isolate Grok OAuth provider adapter
