+++ crates/aether-oauth/src/provider/service.rs
@@ -19,15 +19,17 @@ impl ProviderOAuthService {
     pub fn with_builtin_adapters() -> Self {
         use super::providers::{
             AntigravityProviderOAuthAdapter, CodexProviderOAuthAdapter,
-            GenericProviderOAuthAdapter, KiroProviderOAuthAdapter, WindsurfProviderOAuthAdapter,
+            GenericProviderOAuthAdapter, GrokOAuthProviderOAuthAdapter, KiroProviderOAuthAdapter,
+            WindsurfProviderOAuthAdapter,
         };
 
         let mut service = Self::new()
             .with_adapter(Arc::new(KiroProviderOAuthAdapter::default()))
             .with_adapter(Arc::new(CodexProviderOAuthAdapter::default()))
             .with_adapter(Arc::new(AntigravityProviderOAuthAdapter::default()))
+            .with_adapter(Arc::new(GrokOAuthProviderOAuthAdapter::default()))
             .with_adapter(Arc::new(WindsurfProviderOAuthAdapter));
-        for provider_type in ["claude_code", "chatgpt_web", "gemini_cli", "grok_oauth"] {
+        for provider_type in ["claude_code", "chatgpt_web", "gemini_cli"] {
             if let Some(adapter) = GenericProviderOAuthAdapter::for_provider_type(provider_type) {
                 service = service.with_adapter(Arc::new(adapter));
             }
